extract types to run-call helper
